+++ zeroeq/http/server.h
@@ -6,6 +6,8 @@
 #ifndef ZEROEQ_HTTP_SERVER_H
 #define ZEROEQ_HTTP_SERVER_H
 
+#include "response.h"
+
 #include <zeroeq/http/api.h>
 #include <zeroeq/receiver.h> // base class
 #include <zeroeq/log.h>
@@ -15,6 +17,24 @@ namespace zeroeq
 /** HTTP protocol support. */
 namespace http
 {
+
+/**
+ * HTTP REST callback with payload, returning a Response future.
+ * The string is the request's payload or a query string of a GET request
+ */
+using RESTFunc = std::function< std::future< Response >( const std::string& ) >;
+
+/**
+ * HTTP REST callback for a given path returning a Response future.
+ * The first string provides the url part after the registered endpoint:
+ * "api/windows/jf321f" -> "jf321f".
+ * The second is the request's payload or a query string of a GET request
+ * If the string matches the registered endpoint pass empty string as
+ * path: "api/windows/ -> ""
+ */
+using RESTPathFunc = std::function< std::future< Response >( const std::string&,
+                                                         const std::string& ) >;
+
 /**
  * Serves HTTP GET and PUT requests for servus::Serializable objects.
  *
@@ -91,6 +111,14 @@ public:
      */
     ZEROEQHTTP_API SocketDescriptor getSocketDescriptor() const;
     //@}
+
+    /** Handle a single action on a given endpoint. */
+    bool handle( zeroeq::http::Verb action, const std::string& endpoint,
+                 zeroeq::http::RESTFunc func );
+
+    /** Handle a single action on a given endpoint derived from the path . */
+    bool handlePath( zeroeq::http::Verb action, const std::string& endpoint,
+                     zeroeq::http::RESTPathFunc func );
 
     /** @name Object registration for PUT and GET requests */
     //@{